Read config in agents
