+++ ai_dietolog/agents/intake.py
@@ -10,6 +10,8 @@ from uuid import uuid4
 from typing import Optional
 
 from openai import AsyncOpenAI
+
+from ..core.config import openai_api_key
 from pydantic import ValidationError
 
 from ..core.prompts import MEAL_JSON
@@ -30,7 +32,7 @@ async def intake(
     The meal may include an optional ``clarification`` question generated by the
     AI and stored in ``Meal.clarification``.
     """
-    client = AsyncOpenAI()
+    client = AsyncOpenAI(api_key=openai_api_key())
     system = MEAL_JSON.render(
         meal_type=meal_type,
         user_desc=user_text,
